+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 * [HexaPDF::Rectangle] methods for setting the left, top, right, bottom, width
   and height
 
+### Changed
+
+* All [HexaPDF::Content::ColorSpace] implementations to only normalize values
+  when using the ::color method
+
 
 ## 0.11.9 - 2020-06-15